Fix issue with log files and repo forall command

+++ job.py
@@ -21,9 +21,9 @@ class Job(object):
 
     def add_cmd(self, cmd):
         if self.cmds.empty():
-            log_str =  " 2>&1 | tee %s/%s.txt" % (Job.log_folder, self.id)
+            log_str =  " > %s/%s.txt  2>&1 " % (Job.log_folder, self.id)
         else:
-            log_str =  " 2>&1 | tee -a %s/%s.txt" % (Job.log_folder, self.id)
+            log_str =  " >> %s/%s.txt 2>&1" % (Job.log_folder, self.id)
         self.cmds.put(cmd + log_str)
 
     def build(self):
@@ -77,62 +77,65 @@ class Job(object):
         return
 
 class Qemu(Job):
-    def __init__(self):
+    def __init__(self, id):
         """QEMU class that initialize the job queue"""
-        super().__init__(1, "/home/joakim.bech/devel/optee_projects/qemu")
+        super().__init__(id, "/home/joakim.bech/devel/optee_projects/qemu")
         self.jobtype = "qemu"
         # FIXME: All this could go into "config-files" instead of hardcoding it
         # here.
-        self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
+        if os.path.exists(self.workspace + "/.repo"):
+            self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
         self.add_cmd("repo init -u https://github.com/OP-TEE/manifest.git -m default.xml --reference %s" % Job.reference_folder)
         self.add_cmd("repo sync -j3 -d")
-        self.add_cmd("ln -s %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
+        self.add_cmd("ln -sf %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
         self.add_cmd("cd build && make all -j12")
 
     def __str__(self):
         return "I'm a %s class" % self.jobtype
 
 class Rpi3(Job):
-    def __init__(self):
+    def __init__(self, id):
         """Raspberry Pi3 class that initialize the job queue"""
-        super().__init__(2, "/home/joakim.bech/devel/optee_projects/rpi3")
+        super().__init__(id, "/home/joakim.bech/devel/optee_projects/rpi3")
         self.jobtype = "rpi3"
         # FIXME: All this could go into "config-files" instead of hardcoding it
         # here.
-        self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
+        if os.path.exists(self.workspace + "/.repo"):
+            self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
         self.add_cmd("repo init -u https://github.com/OP-TEE/manifest.git -m %s.xml --reference %s" % (self.jobtype, Job.reference_folder))
         self.add_cmd("repo sync -j3 -d")
-        self.add_cmd("ln -s %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
+        self.add_cmd("ln -sf %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
         self.add_cmd("cd build && make all -j12")
 
     def __str__(self):
         return "I'm a %s class" % self.jobtype
 
 class Fvp(Job):
-    def __init__(self):
+    def __init__(self, id):
         """Foundation Models class that initialize the job queue"""
-        super().__init__(3, "/home/joakim.bech/devel/optee_projects/fvp")
+        super().__init__(id, "/home/joakim.bech/devel/optee_projects/fvp")
         self.jobtype = "fvp"
         # FIXME: All this could go into "config-files" instead of hardcoding it
         # here.
-        self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
+        if os.path.exists(self.workspace + "/.repo"):
+            self.add_cmd("repo forall -c 'git checkout -f && git clean -xdf'")
         self.add_cmd("mkdir -p %s/Foundation_Platformpkg" % self.workspace)
         self.add_cmd("repo init -u https://github.com/OP-TEE/manifest.git -m %s.xml --reference %s" % (self.jobtype, Job.reference_folder))
         self.add_cmd("repo sync -j3 -d")
-        self.add_cmd("ln -s %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
+        self.add_cmd("ln -sf %s/toolchains %s/toolchains" % (Job.reference_folder, self.workspace))
         self.add_cmd("cd build && make all -j12")
 
     def __str__(self):
         return "I'm a %s class" % self.jobtype
 
 def test():
-    #qemu = Qemu()
-    #qemu.build()
+    qemu = Qemu(1)
+    qemu.build()
 
-    #rpi3 = Rpi3()
-    #rpi3.build()
+    rpi3 = Rpi3(2)
+    rpi3.build()
 
-    fvp = Fvp()
+    fvp = Fvp(3)
     fvp.build()
 
 if __name__ == "__main__":
